requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.12
channels~=3.0.4
channels_redis
asgiref~=3.4.1